requirements.txt
pandas
langchain
langchain-community
langchain-groq
faiss-cpu 
sentence-transformers 
streamlit
huggingface_hub
python-dotenv
langchain-huggingface